Move roster table logic into age group section
